requirements.txt
kafka-python
pandas
matplotlib
